fixed setuptools issues with pyproject.toml changes, fixed autodoc issues with conf.py changes, some setup.cfg changes as well, this should resolve all build issues he said as famous last words
