+++ docs/conf.py
@@ -39,7 +39,7 @@ except ImportError:
     from sphinx import apidoc
 
 output_dir = os.path.join(__location__, "api")
-module_dir = os.path.join(__location__, "../roam")
+module_dir = os.path.join(__location__, "../soloroam")
 try:
     shutil.rmtree(output_dir)
 except FileNotFoundError:
@@ -165,7 +165,7 @@ html_theme_options = {
 # The name for this set of Sphinx documents.  If None, it defaults to
 # "<project> v<release> documentation".
 try:
-    from roam import __version__ as version
+    from soloroam import __version__ as version
 except ImportError:
     pass
 else:
@@ -230,7 +230,7 @@ html_static_path = ["_static"]
 # html_file_suffix = None
 
 # Output file base name for HTML help builder.
-htmlhelp_basename = "roam-doc"
+htmlhelp_basename = "soloroam-doc"
 
 
 # -- Options for LaTeX output ------------------------------------------------
