+++ pyproject.toml
@@ -1,17 +1,22 @@
 [tool.poetry]
-name = "roam"
-version = "0.2.1"
+name = "soloroam"
+version = "v0.9.5"
 description = ""
 authors = ["sapporojones <sapporojones@gmail.com>"]
 
 [tool.poetry.dependencies]
 python = "^3.8"
 requests = "^2.25.1"
+setuptools = "^57.1.0"
 
 [tool.poetry.dev-dependencies]
 pytest = "^5.2"
 Sphinx = "^4.0.3"
 
 [build-system]
-requires = ["poetry>=0.12"]
-build-backend = "poetry.masonry.api"
+requires = ["setuptools>=46.1.0", "setuptools_scm[toml]>=5", "wheel"]
+build-backend = "setuptools.build_meta"
+
+[tool.setuptools_scm]
+# See configuration details in https://github.com/pypa/setuptools_scm
+version_scheme = "no-guess-dev"
