+++ setup.cfg
@@ -28,10 +28,10 @@ classifiers =
 
 [options]
 zip_safe = False
-packages = find_namespace:
+packages = find:
 include_package_data = True
 package_dir =
-    =roam
+    =soloroam
 
 # Require a min/specific Python version (comma-separated conditions)
 # python_requires = >=3.8
@@ -41,13 +41,14 @@ package_dir =
 # new major versions. This works if the required packages follow Semantic Versioning.
 # For more information, check out https://semver.org/.
 install_requires =
+    setuptools
     importlib-metadata
     requests
 
 
 
 [options.packages.find]
-where = roam
+where = soloroam
 exclude =
     tests
 
@@ -67,7 +68,7 @@ testing =
 [options.entry_points]
 # Add here console scripts like:
 console_scripts =
-    script_name = roam.run:handler
+    script_name = soloroam.run:handler
 # For example:
 # console_scripts =
 #     fibonacci = eyeson_flask.skeleton:run
@@ -90,7 +91,7 @@ norecursedirs =
     .tox
     .idea
 
-testpaths = roam/tests
+testpaths = soloroam/tests
 # Use pytest markers to select/deselect specific tests
 # markers =
 #     slow: mark tests as slow (deselect with '-m "not slow"')
