+++ soloroam/encounters.py
