+++ tox.ini
@@ -48,7 +48,6 @@ setenv =
     docs: BUILD = html
     doctests: BUILD = doctest
 deps =
-    pytest
     -r {toxinidir}/requirements.txt
     # ^  requirements.txt shared with Read The Docs
 commands =
